Fix Oracle Logging CI test isolation
Share the mock logging client across logging specs so tests that enable OCI logging do not instantiate a real OCI SDK client, and reset shared Logback appender state before integration assertions.
