+++ oraclecloud-logging/src/test/groovy/io/micronaut/oraclecloud/logging/OracleCloudLoggingSpec.groovy
@@ -3,23 +3,16 @@ package io.micronaut.oraclecloud.logging
 import ch.qos.logback.classic.Logger
 import ch.qos.logback.classic.LoggerContext
 import ch.qos.logback.core.read.ListAppender
-import com.oracle.bmc.Region
 import com.oracle.bmc.loggingingestion.Logging
-import com.oracle.bmc.loggingingestion.LoggingClient
 import com.oracle.bmc.loggingingestion.model.LogEntry
 import com.oracle.bmc.loggingingestion.model.LogEntryBatch
-import com.oracle.bmc.loggingingestion.requests.PutLogsRequest
-import com.oracle.bmc.loggingingestion.responses.PutLogsResponse
 import io.micronaut.context.annotation.Property
-import io.micronaut.context.annotation.Replaces
-import io.micronaut.context.annotation.Requires
 import io.micronaut.context.event.ApplicationEventPublisher
 import io.micronaut.runtime.ApplicationConfiguration
 import io.micronaut.runtime.server.EmbeddedServer
 import io.micronaut.runtime.server.event.ServerStartupEvent
 import io.micronaut.test.extensions.spock.annotation.MicronautTest
 import jakarta.inject.Inject
-import jakarta.inject.Singleton
 import org.slf4j.LoggerFactory
 import spock.lang.Specification
 import spock.util.concurrent.PollingConditions
@@ -47,17 +40,24 @@ class OracleCloudLoggingSpec extends Specification {
         def logMessage = 'test logging'
         def testHost = 'testHost'
         def logger = LoggerFactory.getLogger(OracleCloudLoggingSpec.class)
+        def contextLogger = LoggerFactory.getLogger('io.micronaut.context.env.DefaultEnvironment')
         PollingConditions conditions = new PollingConditions(timeout: 10, initialDelay: 1.5, factor: 1.25)
         LoggerContext loggerContext = (LoggerContext) LoggerFactory.getILoggerFactory()
         ListAppender listAppender
+        OracleCloudAppender oracleCloudAppender
         loggerContext.loggerList.each { Logger l ->
             l.iteratorForAppenders().each { appender ->
                 if (appender.name == 'ORACLE') {
-                    OracleCloudAppender oracleCloudAppender = (OracleCloudAppender) appender
+                    oracleCloudAppender = (OracleCloudAppender) appender
                     listAppender = (ListAppender) oracleCloudAppender.getAppender('MOCK')
                 }
             }
         }
+        oracleCloudAppender.@configuredSuccessfully = false
+        oracleCloudAppender.source = null
+        oracleCloudAppender.type = null
+        oracleCloudAppender.subject = null
+        listAppender.list.clear()
 
         when:
         def instance = Mock(EmbeddedServer.class)
@@ -65,6 +65,7 @@ class OracleCloudLoggingSpec extends Specification {
         def mockLogging = (MockLogging) logging
         1 * instance.getHost() >> testHost
         eventPublisher.publishEvent(event)
+        contextLogger.info('Established active environments: [oraclecloud]')
         logger.info(logMessage)
 
         then:
@@ -84,7 +85,8 @@ class OracleCloudLoggingSpec extends Specification {
                 }
         )
         logEntryBatch.stream().allMatch(x -> x.source == testHost)
-        logEntryBatch.stream().allMatch(x -> x.type == testHost + '.' + applicationConfiguration.getName().get())
+        String expectedType = "${testHost}.${applicationConfiguration.getName().get()}"
+        logEntryBatch.stream().allMatch(x -> expectedType.equals(x.type))
         logEntryBatch.stream().anyMatch(x -> x.subject == applicationConfiguration.getName().get())
 
         logEntries.stream().anyMatch(x -> x.data.contains('io.micronaut.context'))
@@ -105,76 +107,6 @@ class OracleCloudLoggingSpec extends Specification {
             listAppender.list.size() != 0
         }
         listAppender.list.get(0).message == logMessage
-    }
-
-    @Requires(property = "spec.name", value = "OracleCloudLoggingSpec")
-    @Singleton
-    @Replaces(LoggingClient)
-    static class MockLogging implements Logging {
-
-        final List<PutLogsRequest> putLogsRequestList = Collections.synchronizedList(new ArrayList<>())
-
-        private boolean success = true
-
-        @Override
-        void refreshClient() {
-
-        }
-
-        @Override
-        void setEndpoint(String endpoint) {
-
-        }
-
-        @Override
-        void useRealmSpecificEndpointTemplate(boolean use) {
-
-        }
-
-        boolean getSuccess() {
-            return success
-        }
-
-        void setSuccess(boolean success) {
-            this.success = success
-        }
-
-        @Override
-        String getEndpoint() {
-            return 'mock-logging-endpoint'
-        }
-
-        @Override
-        void setRegion(Region region) {
-
-        }
-
-        @Override
-        void setRegion(String regionId) {
-
-        }
-
-        List<PutLogsRequest> getPutLogsRequestList() {
-            synchronized (putLogsRequestList) {
-                return new ArrayList<PutLogsRequest>(putLogsRequestList)
-            }
-        }
-
-        @Override
-        PutLogsResponse putLogs(PutLogsRequest request) {
-            synchronized (putLogsRequestList) {
-                putLogsRequestList.add(request)
-            }
-            if (success) {
-                return PutLogsResponse.builder().opcRequestId("validId").build()
-            }
-            return PutLogsResponse.builder().opcRequestId(null).__httpStatusCode__(404).build()
-        }
-
-        @Override
-        void close() throws Exception {
-
-        }
     }
 
 }
